source/requirements.txt
altgraph                  0.17.4
Jinja2                    3.1.4
MarkupSafe                3.0.2
packaging                 24.1
pefile                    2023.2.7
pip                       24.2
pycryptodomex             3.21.0
pyinstaller               6.9.0
pyinstaller-hooks-contrib 2024.8
PySide6                   6.7.2
PySide6_Addons            6.7.2
PySide6_Essentials        6.7.2
pywin32-ctypes            0.2.2
QDarkStyle                3.2.3
qt-material               2.14
QtPy                      2.4.2
setuptools                65.5.0
shiboken6                 6.7.2